Refactoring to move the query manager to its own module
